all types changes do int

+++ src/main/java/Calculator.java
@@ -1,13 +1,13 @@
 public class Calculator {
-    public double add (double a, double b){
+    public int add (int a, int b){
         return a + b;
     }
 
-    public double multiply (double a, double b){
+    public int multiply (int a, int b){
         return a * b;
     }
 
-    public double divide (double a, double b){
+    public int divide (int a, int b){
         if (b == 0.0) {
             throw new ArithmeticException("Dividing by 0 is not possible");
         }
